refactoring addMarkup function
