--- html/src/js/advantages/addMarkup.js
@@ -1,12 +1,0 @@
-function addMarkup(ref, tmp, data) {
-  Handlebars.registerHelper('alternatingClass', function (index, class1, class2) {
-  return (index % 2 !== 0) ? class1 : class2;
-});
-
-  const template = Handlebars.compile(tmp);
-  const markup = template(data);
-
-  ref.innerHTML = markup;
-}
-
-export default addMarkup;
