+++ html/src/js/common/addMarkupFromTemplate.js
