+++ html/src/js/problems/addImgContentToCards.js
@@ -1,4 +1,4 @@
-import addMarkup from "./addMarkup.js";
+import addMarkup from '../common/addMarkupFromTemplate.js';
 import loadTemplate from "../common/loadTemplate.js";
 
 export async function addImgContentToCard(refsArr, templatesLinksArr) {
@@ -7,7 +7,7 @@ export async function addImgContentToCard(refsArr, templatesLinksArr) {
   templatesLinksArr.forEach((tmpLink, i) => {
     try {
       loadTemplate(tmpLink).then(tmp => {
-        addMarkup(refsArr[i], tmp, []);
+      addMarkup(refsArr[i], tmp, []);
       });
     } catch (err) {
       throw new Error(err)
